move config folder into src and change routes of keys and passport in backend
