+++ frontend/src/config/keys_prod.js
@@ -2,4 +2,5 @@ module.exports = {
   mongoURI: process.env.MONGO_URI,
   secretToken: process.env.SECRET_TOKEN,
   yelpAPIKey: process.env.YELP_API_KEY,
+  googleMapsAPIKey: process.env.GOOGLE_MAPS_API_KEY
 }
