+++ routes/api/users.js
@@ -2,7 +2,7 @@ const express = require("express");
 const bcrypt = require("bcryptjs");
 const jwt = require('jsonwebtoken'); // for our session tokens
 const passport = require('passport');
-const keys = require('../../config/keys');
+const keys = require('../../frontend/src/config/keys');
 const router = express.Router();
 const User = require("../../models/User"); // get User mongoose object
 const validateSignUpInput = require('../../validation/signup');
